// src/api/generated/model/task.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { Board } from './board';
import type { TaskStatus } from './taskStatus';

export interface Task {
  id?: number;
  title?: string;
  description?: string;
  status?: TaskStatus;
  board?: Board;
}