// src/api/generated/model/taskRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface TaskRequest {
  /**
     * @minLength 0
     * @maxLength 100
     */
  title: string;
  /**
     * @minLength 0
     * @maxLength 500
     */
  description?: string;
  boardId: number;
  status?: string;
}